chore: PSR-12 formatting + safe Rector cleanup (#99)
* style: apply PSR-12 formatting across app, tests and public entrypoints

Whitespace/structure-only pass (blank line after the opening tag, brace
position, statement indentation, import cleanup, operator spacing). No
behavioural changes.

Validated: full PHPUnit suite green (84 tests); PHPStan delta is zero
(the pre-existing baseline-uncovered findings are unchanged).

* refactor: remove redundant casts (Rector, strictly safe subset)

Behaviour-preserving cleanup only, hand-picked rules (15 files):
- RecastingRemovalRector — drop casts already matching the value's type
- RemoveConcatAutocastRector — drop (string) casts redundant in concatenation
- RemoveUnusedNonEmptyArrayBeforeForeachRector — drop redundant if-guard before foreach

RemoveUnusedVariableAssignRector was evaluated and REJECTED: it dropped the
assignment of a security filter's return value in PageController
(`$relatedAlbums = $this->filterAlbumsByAccess(...)` -> bare call, which
discards the filtered result and would leak private/NSFW related albums), so
it is excluded. The broader Rector sets (readonly, switch->match, explicit-bool
compare, first-class-callable, constructor promotion) remain deferred to a
separately-reviewed pass.

Validated: PHPUnit 84 green; PHPStan delta zero; style re-checked with the
formatter.

+++ app/Controllers/Admin/CollectionsController.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace App\Controllers\Admin;
@@ -310,7 +311,7 @@ final class CollectionsController extends BaseController
             'total' => $total,
             'page' => $page,
             'per_page' => self::PICKER_PER_PAGE,
-            'total_pages' => (int) max(1, (int) ceil($total / self::PICKER_PER_PAGE)),
+            'total_pages' => max(1, (int) ceil($total / self::PICKER_PER_PAGE)),
         ]);
     }
